+++ lib/binance_client/requests/authenticated_base_request.rb
@@ -2,11 +2,13 @@ module BinanceClient
   class AuthenticatedBaseRequest < BaseRequest
     include APIClientBase::Request.module
 
+    attribute :recv_window, Integer
+
     def signature
       OpenSSL::HMAC.hexdigest(
         OpenSSL::Digest.new("sha256"),
         api_secret,
-        params_without_signature_with_timestamp.to_query,
+        params_without_signature_with_timestamp_and_recv_window.to_query,
       )
     end
 
@@ -19,11 +21,14 @@ module BinanceClient
     end
 
     def params
-      params_without_signature_with_timestamp.merge(signature_hash).to_query
+      params_without_signature_with_timestamp_and_recv_window.
+        merge(signature_hash).to_query
     end
 
-    def params_without_signature_with_timestamp
-      params_without_signature.merge(timestamp: timestamp)
+    def params_without_signature_with_timestamp_and_recv_window
+      attrs_to_merge = { timestamp: timestamp }
+      attrs_to_merge[:recvWindow] = recv_window if recv_window.present?
+      params_without_signature.merge(attrs_to_merge)
     end
 
     def params_without_signature