+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove `used_weight` from responses
 
 ### Added
+- `recv_window` configuration to set `recvWindow`
+- `#account` call to get the `/api/v3/account` info
+
+### Added
 - `used_weights` to responses
 
 ## [1.4.0] - 2021-12-09